Refactor match-filter
