+++ eqcorrscan/core/__init__.py
@@ -12,7 +12,7 @@ import importlib
 import warnings
 
 __all__ = ['template_gen', 'match_filter', 'bright_lights', 'lag_calc',
-           'subspace', 'subspace_statistic']
+           'subspace']
 
 _import_map = {}
