+++ eqcorrscan/core/match_filter/detection.py
@@ -1,0 +1,350 @@
+"""
+Functions for network matched-filter detection of seismic data.
+
+Designed to cross-correlate templates generated by template_gen function
+with data and output the detections.
+
+:copyright:
+    EQcorrscan developers.
+
+:license:
+    GNU Lesser General Public License, Version 3
+    (https://www.gnu.org/copyleft/lesser.html)
+"""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import ast
+import copy
+import os
+import logging
+
+import numpy as np
+from obspy import Catalog, UTCDateTime
+from obspy.core.event import (
+    Comment, WaveformStreamID, Event, Pick, CreationInfo, ResourceIdentifier)
+
+from eqcorrscan.core.match_filter.helpers import _test_event_similarity
+
+Logger = logging.getLogger(__name__)
+
+
+class Detection(object):
+    """
+    Single detection from detection routines in eqcorrscan.
+    Information required for a full detection based on cross-channel \
+    correlation sums.
+
+    :type template_name: str
+    :param template_name: The name of the template for which this \
+        detection was made.
+    :type detect_time: obspy.core.utcdatetime.UTCDateTime
+    :param detect_time: Time of detection as an obspy UTCDateTime object
+    :type no_chans: int
+    :param no_chans: The number of channels for which the cross-channel \
+        correlation sum was calculated over.
+    :type detect_val: float
+    :param detect_val: The raw value of the cross-channel correlation sum \
+        for this detection.
+    :type threshold: float
+    :param threshold: The value of the threshold used for this detection, \
+        will be the raw threshold value related to the cccsum.
+    :type typeofdet: str
+    :param typeofdet: Type of detection, STA, corr, bright
+    :type threshold_type: str
+    :param threshold_type: Type of threshold used for detection
+    :type threshold_input: float
+    :param threshold_input:
+        Threshold set for detection, relates to `threshold` according to the
+        `threshold_type`.
+    :type chans: list
+    :param chans: List of stations for the detection
+    :type event: obspy.core.event.event.Event
+    :param event:
+        Obspy Event object for this detection, note that this is lost when
+        writing to a :class:`Detection` objects to csv files using
+        :func:`eqcorrscan.core.match_filter.Detection.write`
+    :type id: str
+    :param id: Identification for detection (should be unique).
+    """
+
+    def __init__(self, template_name, detect_time, no_chans, detect_val,
+                 threshold, typeofdet, threshold_type, threshold_input,
+                 chans=None, event=None, id=None):
+        """Main class of Detection."""
+        self.template_name = template_name
+        self.detect_time = detect_time
+        self.no_chans = int(no_chans)
+        if not isinstance(chans, list):
+            self.chans = [chans]
+        else:
+            self.chans = chans
+        self.detect_val = np.float32(detect_val)
+        self.threshold = np.float32(threshold)
+        self.typeofdet = typeofdet
+        self.threshold_type = threshold_type
+        self.threshold_input = threshold_input
+        self.event = event
+        if id is not None:
+            self.id = id
+        else:
+            self.id = (''.join(template_name.split(' ')) + '_' +
+                       detect_time.strftime('%Y%m%d_%H%M%S%f'))
+        if event is not None:
+            event.resource_id = self.id
+        if self.typeofdet == 'corr':
+            assert abs(self.detect_val) <= self.no_chans
+
+    def __repr__(self):
+        """Simple print."""
+        print_str = ' '.join(
+            ['template name =', self.template_name, '\n',
+             'detection id =', self.id, '\n',
+             'detection time =', str(self.detect_time), '\n',
+             'number of channels =', str(self.no_chans), '\n',
+             'channels =', str(self.chans), '\n',
+             'detection value =', str(self.detect_val), '\n',
+             'threshold =', str(self.threshold), '\n',
+             'threshold type =', self.threshold_type, '\n',
+             'input threshold =', str(self.threshold_input), '\n',
+             'detection type =', str(self.typeofdet)])
+        return "Detection(" + print_str + ")"
+
+    def __str__(self):
+        """Full print."""
+        return (
+            "Detection on template: {0} at: {1} with {2} channels: {3}".format(
+                self.template_name, self.detect_time, self.no_chans,
+                self.chans))
+
+    def __eq__(self, other, verbose=False):
+        for key in self.__dict__.keys():
+            self_is_event = isinstance(self.event, Event)
+            other_is_event = isinstance(other.event, Event)
+            if key == 'event':
+                if self_is_event and other_is_event:
+                    if not _test_event_similarity(
+                            self.event, other.event, verbose=verbose):
+                        return False
+                elif self_is_event and not other_is_event:
+                    return False
+                elif not self_is_event and other_is_event:
+                    return False
+            elif self.__dict__[key] != other.__dict__[key]:
+                return False
+        return True
+
+    def __lt__(self, other):
+        if self.detect_time < other.detect_time:
+            return True
+        else:
+            return False
+
+    def __le__(self, other):
+        if self.detect_time <= other.detect_time:
+            return True
+        else:
+            return False
+
+    def __gt__(self, other):
+        return not self.__le__(other)
+
+    def __ge__(self, other):
+        return not self.__lt__(other)
+
+    def __hash__(self):
+        """
+        Cannot hash Detection objects, they may change.
+        :return: 0
+        """
+        return 0
+
+    def __ne__(self, other):
+        return not self.__eq__(other)
+
+    def copy(self):
+        """
+        Returns a copy of the detection.
+
+        :return: Copy of detection
+        """
+        return copy.deepcopy(self)
+
+    def write(self, fname, append=True):
+        """
+        Write detection to csv formatted file.
+
+        Will append if append==True and file exists
+
+        :type fname: str
+        :param fname: Full path to file to open and write to.
+        :type append: bool
+        :param append:
+            Set to true to append to an existing file, if True and file doesn't
+            exist, will create new file and warn.  If False will overwrite
+            old files.
+        """
+        mode = 'w'
+        if append and os.path.isfile(fname):
+            mode = 'a'
+        header = '; '.join([
+            'Template name', 'Detection time (UTC)', 'Number of channels',
+            'Channel list', 'Detection value', 'Threshold', 'Threshold type',
+            'Input threshold', 'Detection type'])
+        print_str = "{0}; {1}; {2}; {3}; {4}; {5}; {6}; {7}; {8}\n".format(
+            self.template_name, self.detect_time, self.no_chans,
+            self.chans, self.detect_val, self.threshold,
+            self.threshold_type, self.threshold_input, self.typeofdet)
+        with open(fname, mode) as _f:
+            _f.write(header + '\n')  # Write a header for the file
+            _f.write(print_str)
+
+    def _calculate_event(self, template=None, template_st=None):
+        """
+        Calculate an event for this detection using a given template.
+
+        :type template: Template
+        :param template: The template that made this detection
+        :type template_st: `obspy.core.stream.Stream`
+        :param template_st:
+            Template stream, used to calculate pick times, not needed if
+            template is given.
+
+        .. rubric:: Note
+            Works in place on Detection - over-writes previous events.
+            Does not correct for pre-pick.
+        """
+        if template is not None and template.name != self.template_name:
+            Logger.info("Template names do not match: {0}: {1}".format(
+                template.name, self.template_name))
+            return
+        # Detect time must be valid QuakeML uri within resource_id.
+        # This will write a formatted string which is still
+        # readable by UTCDateTime
+        det_time = str(self.detect_time.strftime('%Y%m%dT%H%M%S.%f'))
+        ev = Event(resource_id=ResourceIdentifier(
+            id=self.template_name + '_' + det_time,
+            prefix='smi:local'))
+        ev.creation_info = CreationInfo(
+            author='EQcorrscan', creation_time=UTCDateTime())
+        ev.comments.append(
+            Comment(text='threshold={0}'.format(self.threshold)))
+        ev.comments.append(
+            Comment(text='detect_val={0}'.format(self.detect_val)))
+        if self.chans is not None:
+            ev.comments.append(
+                Comment(text='channels used: {0}'.format(
+                    ' '.join([str(pair) for pair in self.chans]))))
+        if template is not None:
+            template_st = template.st
+        min_template_tm = min(
+            [tr.stats.starttime for tr in template_st])
+        for tr in template_st:
+            if (tr.stats.station, tr.stats.channel) \
+                    not in self.chans:
+                continue
+            elif tr.stats.__contains__("not_in_original"):
+                continue
+            else:
+                pick_time = self.detect_time + (
+                        tr.stats.starttime - min_template_tm)
+                ev.picks.append(Pick(
+                    time=pick_time, waveform_id=WaveformStreamID(
+                        network_code=tr.stats.network,
+                        station_code=tr.stats.station,
+                        channel_code=tr.stats.channel,
+                        location_code=tr.stats.location)))
+        self.event = ev
+        return
+
+
+def read_detections(fname, encoding="UTF8"):
+    """
+    Read detections from a file to a list of Detection objects.
+
+    :type fname: str
+    :param fname:
+        File to read from, must be a file written to by Detection.write.
+    :type encoding: str
+    :param encoding: Encoding used for detection file.
+
+    :returns: list of :class:`eqcorrscan.core.match_filter.Detection`
+    :rtype: list
+
+    .. note::
+        :class:`eqcorrscan.core.match_filter.detection.Detection`'s returned
+        do not contain Detection.event
+    """
+    with open(fname, "rb") as _f:
+        lines = _f.read().decode(encoding).splitlines()
+    detections = []
+    for index, line in enumerate(lines):
+        if index == 0:
+            continue  # Skip header
+        if line.rstrip().split('; ')[0] == 'Template name':
+            continue  # Skip any repeated headers
+        detection = line.rstrip().split('; ')
+        detection[1] = UTCDateTime(detection[1])
+        detection[2] = int(float(detection[2]))
+        detection[3] = ast.literal_eval(detection[3])
+        detection[4] = float(detection[4])
+        detection[5] = float(detection[5])
+        if len(detection) < 9:
+            detection.extend(['Unset', float('NaN')])
+        else:
+            detection[7] = float(detection[7])
+        detections.append(Detection(
+            template_name=detection[0], detect_time=detection[1],
+            no_chans=detection[2], detect_val=detection[4],
+            threshold=detection[5], threshold_type=detection[6],
+            threshold_input=detection[7], typeofdet=detection[8],
+            chans=detection[3]))
+    return detections
+
+
+def write_catalog(detections, fname, format="QUAKEML"):
+    """Write events contained within detections to a catalog file.
+
+    :type detections: list
+    :param detections: list of eqcorrscan.core.match_filter.Detection
+    :type fname: str
+    :param fname: Name of the file to write to
+    :type format: str
+    :param format: File format to use, see obspy.core.event.Catalog.write \
+        for supported formats.
+    """
+    catalog = get_catalog(detections)
+    catalog.write(filename=fname, format=format)
+
+
+def get_catalog(detections):
+    """
+    Generate an :class:`obspy.core.event.Catalog` from list of \
+    :class:`Detection`'s.
+
+    :type detections: list
+    :param detections: list of :class:`eqcorrscan.core.match_filter.Detection`
+
+    :returns: Catalog of detected events.
+    :rtype: :class:`obspy.core.event.Catalog`
+
+    .. warning::
+        Will only work if the detections have an event associated with them.
+        This will not be the case if detections have been written to csv
+        format using :func:`eqcorrscan.core.match_filter.Detection.write`
+        and read back in.
+    """
+    catalog = Catalog()
+    for detection in detections:
+        if detection.event:
+            catalog.append(detection.event)
+    return catalog
+
+
+if __name__ == "__main__":
+    import doctest
+
+    doctest.testmod()
