+++ eqcorrscan/utils/sac_util.py
@@ -132,19 +132,19 @@ def sactoevent(st):
                 if tr.stats.sac[pick_key] == float_nan:
                     # in version 0.10.2 and before. rather than not include
                     # the keys, the variables are filled with SAC nans.
-                    Logger.warning(
+                    Logger.debug(
                         'No pick in position {0} for trace {1}'.format(
                             pick_key, tr.id))
                     continue
                 pick_time = reference_time + tr.stats.sac[pick_key]
                 phase_hint = tr.stats.sac[phase_key].split()[0]
             except KeyError:
-                Logger.warning(
-                        'No pick in position {0} for trace {1}'.format(
-                            pick_key, tr.id))
+                Logger.debug(
+                    'No pick in position {0} for trace {1}'.format(
+                        pick_key, tr.id))
                 continue
-            Logger.info('Found pick in position {0} for {1}'.format(
-                pick_key, tr.id))
+            Logger.info(
+                'Found pick in position {0} for {1}'.format(pick_key, tr.id))
             waveform_id = WaveformStreamID(station_code=tr.stats.station,
                                            network_code=tr.stats.network,
                                            channel_code=tr.stats.channel)
@@ -155,19 +155,18 @@ def sactoevent(st):
         # Also check header slots 'a' and 'ka'
         try:
             if tr.stats.sac['a'] == float_nan:
-                Logger.warning(
-                        'No pick in position {0} for trace {1}'.format(
-                            pick_key, tr.id))
+                Logger.debug(
+                    'No pick in position {0} for trace {1}'.format(
+                        pick_key, tr.id))
                 continue
             pick_time = reference_time + tr.stats.sac['a']
             phase_hint = tr.stats.sac['ka'].split()[0]
         except KeyError:
-            Logger.warning(
-                        'No pick in position {0} for trace {1}'.format(
-                            pick_key, tr.id))
+            Logger.debug(
+                'No pick in position {0} for trace {1}'.format(pick_key, tr.id))
             continue
-        Logger.info('Found pick in position {0} for {1}'.format(
-            pick_key, tr.id))
+        Logger.info(
+            'Found pick in position {0} for {1}'.format(pick_key, tr.id))
         waveform_id = WaveformStreamID(
             station_code=tr.stats.station, network_code=tr.stats.network,
             channel_code=tr.stats.channel)
